fix: add README and delete unsused models
