+++ src/RecipesFinder_bot/Resources/Ingredient.resx
@@ -117,10 +117,13 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="Link" xml:space="preserve">
-    <value>/ingredient</value>
+  <data name="Exception" xml:space="preserve">
+    <value>Hmm.. This is not a ingredient, try sending something else!</value>
   </data>
   <data name="Message" xml:space="preserve">
-    <value />
+    <value>Keep a recipes: {0}</value>
+  </data>
+  <data name="Text" xml:space="preserve">
+    <value>/ingredient</value>
   </data>
 </root>
