+++ src/RecipesFinder_bot/appsettings.json
@@ -7,7 +7,7 @@
     }
   },
   "AllowedHosts": "*",
-  "Url": "https://recipesfinder-bot.herokuapp.com/",
+  "Url": "https://recipesfinder-bot.herokuapp.com",
   "Token": "1148344323:AAEjZNPojEnCgeVdV4dBUO3ipXpKGDpwuMM",
   "Api": {
     "api_key": "d67605a945664091b46863dfb731c31a" //ключ Влада
